Performing ftrace hooks
This commit performs the hooking. The main idea is to change the %rip
register in the ftrace function so that it points to the entry of the
hook function. After we exit the ftrace function the CPU will execute
the hook function instead of the original one.

The hook function can still execute the original function (using the
pointer saved beforehand). However, this means that the ftrace function
will be executed again.

To break the infinite recursion loop we use explicit task list. When
a task executes a hooked function the ftrace function will make a note
of the task and will not switch call the hook function from the nested
call of the original function.

Note that this approach has some limitations:

- We cannot hook truly recursive functions. Thankfully, most of the
  kernel functions are not recursive so this is fine.

- The task list has fixed size. We should not perform any allocations
  in the ftrace function, so we limit the number of concurrent
  activations of the hooked function in the system. Note that this
  is *not* the number of CPUs: preemptive kernel can have arbitrary
  number of suspended tasks which are currently executing the hooked
  function.

- If the hook function does not call the original function then the
  task must be manually removed from the active task list. Otherwise
  the next execution of the hooked function will not be hooked.