Simplify recursion detection
We do not actually need to track calling tasks to avoid vicious loop of
hook calls. We can check the parent_ip passed to the ftrace callback. If
the address lies without our module then we know that it's a call to the
original function from the hook.  If that's the case then we should not
modify the %rip register and allow the original function to execute.