Introduce alternative to return-address checks
There is an alternative way of avoiding vicious recursion during
original function call: jump over the fentry call. Add it to the code
base and allow switching between these options.

Also add a pragma to disable GCC's optimization which may interfere
with return-address checking. This code does not trigger the issue, but
it will be there if pr_info() are replaced with pr_devel() in the hook
functions.