Support newer 4.17+ kernels for real (#5)
Previously mentioned patchset changed not only the naming convention of
x86_64 stubs but their ABI as well. Now they receive a pointer to struct
pt_regs laid out on the stack. We have to pass this pointer as is to the
original handler, and if we wish to examine the arguments of a system
call then we have to look into appropriate registers ourselves.